move test files so `publish` does not include them
