+++ test/.eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "extends": [
+    "../.eslintrc.json"
+  ],
+  "rules": {
+    "node/no-unpublished-require": 0
+  }
+}
